\name{moult-package}
\alias{moult-package}
\docType{package}
\title{Methods for Estimating Start and Duration of Moult in Birds}
\description{Functions for estimating the duration and mean start date for avian moult data, based on maximum likelihood}
\details{
\tabular{ll}{
Package: \tab moult\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-03-04\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\author{Birgit Erni}

\references{Underhill, L. G. and Zucchini, W. (1998). A model for avian primary moult. Ibis 130: 358-372.

Underhill, L. G. and Zucchini, W. and Summers, R. W. (1990). A model for avian primary moult-data types based on migration strategies and an example using the Redshank \emph{Tringa totanus}. Ibis 132: 118-123. }

\keyword{ package }

