\name{moult-package}
\alias{moult-package}
\docType{package}
\title{Methods for Estimating Start and Duration of Moult in Birds}
\description{Functions for estimating the duration and mean start date
  for avian moult data, based on maximum likelihood.}
\details{
\tabular{ll}{
Package: \tab moult\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2012-06-07\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\author{Birgit Erni}

\references{Erni, B., Bonnevie, B. T., Oschadleus, H.-D., Altwegg,
        R. and Underhill, L. G. (2013) moult: An R package to analyze
        moult in birds. Journal of Statistical Software, \bold{52}(8), 1--23.
        \url{http://www.jstatsoft.org/v52/i08/}

  Underhill, L. G. and Zucchini, W. (1988) A model for avian
  primary moult. Ibis \bold{130}, 358--372.

  Underhill, L. G. and Zucchini, W. and Summers, R. W. (1990) A model
  for avian primary moult-data types based on migration strategies and
  an example using the Redshank \emph{Tringa totanus}. Ibis \bold{132},
  118--123.}

\keyword{package}

